feat(error): implement error handling in service and viewmodel

+++ TravellingMap/DataServices/LocationsDataService.swift
@@ -22,39 +22,38 @@ class LocationsDataService {
         return Date().timeIntervalSince(lastFetchDate) > refreshInterval
     }
     
-    func fetchAndSaveData(context: ModelContext) async {
+    func fetchAndSaveData(context: ModelContext) async throws {
         guard shouldFetchData() else { return }
-        do {
-            guard let url = URL(string: "https://gist.githubusercontent.com/danghoang03/c499b00fb63cbf76cf32ac94dcc85d24/raw/5c62ec40311419ee5a14b92047c6d07369085596/LocationsData.json") else {
-                print("URL Error")
-                return
-            }
-            
-            let (data,_) = try await URLSession.shared.data(from: url)
-            let locationsDTO = try JSONDecoder().decode([LocationDTO].self, from: data)
-            
-            let receivedIDs = locationsDTO.map { $0.name + $0.cityName }
-            
-            let favoriteDescriptor = FetchDescriptor<Location>(predicate: #Predicate { $0.isFavorite })
-            let existingFavorites = try? context.fetch(favoriteDescriptor)
-            let favoriteIDs = Set(existingFavorites?.map { $0.id } ?? [])
-            
-            try? context.delete(model: Location.self, where: #Predicate { location in
-                !receivedIDs.contains(location.id)
-            })
-            
-            for locationDTO in locationsDTO {
-                let locationModel = locationDTO.toModel()
-                if favoriteIDs.contains(locationModel.id) {
-                    locationModel.isFavorite = true
-                }
-                context.insert(locationModel)
-            }
-            try context.save()
-            UserDefaults.standard.set(Date(), forKey: "lastFetchDate")
-        } catch {
-            print("Error fetching data: \(error.localizedDescription)")
+        guard let url = URL(string: "https://gist.githubusercontent.com/danghoang03/c499b00fb63cbf76cf32ac94dcc85d24/raw/5c62ec40311419ee5a14b92047c6d07369085596/LocationsData.json") else {
+            throw URLError(.badURL)
         }
+            
+        let (data,response) = try await URLSession.shared.data(from: url)
+        guard let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200 else {
+            throw URLError(.badServerResponse)
+        }
+            
+        let locationsDTO = try JSONDecoder().decode([LocationDTO].self, from: data)
+        
+        let receivedIDs = locationsDTO.map { $0.name + $0.cityName }
+            
+        let favoriteDescriptor = FetchDescriptor<Location>(predicate: #Predicate { $0.isFavorite })
+        let existingFavorites = try? context.fetch(favoriteDescriptor)
+        let favoriteIDs = Set(existingFavorites?.map { $0.id } ?? [])
+            
+        try? context.delete(model: Location.self, where: #Predicate { location in
+            !receivedIDs.contains(location.id)
+        })
+        
+        for locationDTO in locationsDTO {
+            let locationModel = locationDTO.toModel()
+            if favoriteIDs.contains(locationModel.id) {
+                locationModel.isFavorite = true
+            }
+            context.insert(locationModel)
+        }
+        try context.save()
+        UserDefaults.standard.set(Date(), forKey: "lastFetchDate")
     }
     
     func forceRefresh() {
